+++ module/dice/dhRolls.mjs
